feat: funcionando

+++ TP1 - Conversor de Expressões/include/item.hpp
@@ -6,13 +6,13 @@ public:
 
     enum class TipoOperador { Soma, Subtracao, Multiplicacao, Divisao, Numero }
 
-    double valor;
+    long double valor;
     TipoOperador operador;
 
     bool isOperador();
     bool isValue();
 
-    Item(double valor, TipoOperador operador);
+    Item(long double valor, TipoOperador operador);
     ~Item();
 };
